Removes changes unrelated to previous commit message
The latest real commit on master bd5bd26f6ae2c4402403bd47d8258bc6d209e334
contains the message "Removed the Password from the file". I assume this
commit was supposed to only modify the `src/main/resources/application-mlab.yaml`
and not include the other 4 files that seem to include an incomplete
test case that causes the build to fail.

This failing test seems to be the cause of the greenkeeper builds failing
which resulted in greenkeeper opening the issues #44, #45, #48, #49, #52
and #58.

Disclaimer: I haven't gone through the build output of all builds but
the ones I did look at failed on this test or a timeout downloading the
mongodb runtime.

Removing a test instead of fixing the test isn't usually my style but in
this case it appears that the code should not have been comitted in the
first place.